can now add/delete teams and team seekers from session

+++ JamTeamFormingTool/Services/JamTeamFormingSessionService.cs
@@ -99,6 +99,16 @@ namespace JamTeamFormingTool.Services
             {
                 return false;
             }
+            List<Participant> participants = _dbContext.Participants.Where(p => p.Session == maybeSession).ToList();
+            foreach (Participant participant in participants)
+            {
+                DeleteParticipantByID(participant.ID);
+            }
+            List<Team> teams = _dbContext.Teams.Where(t => t.Session == maybeSession).ToList();
+            foreach (Team team in teams)
+            {
+                DeleteTeamByID(team.ID);
+            }
             _dbContext.JamTeamFormingSessions.Remove(maybeSession);
             _dbContext.SaveChanges();
             return true;
@@ -212,7 +222,8 @@ namespace JamTeamFormingTool.Services
         public List<Participant> GetParticipantsForSession(JamTeamFormingSession session)
         {
             return _dbContext.Participants
-                .Where(t => t.Session == session)
+                .Where(p => p.Session == session)
+                .Include(p => p.Roles.OrderBy(r => r.Name))
                 .OrderBy(p => p.Name)
                 .AsNoTracking()
                 .ToList();
@@ -222,30 +233,83 @@ namespace JamTeamFormingTool.Services
         {
             return _dbContext.Teams
                 .Where(t => t.Session == session)
+                .Include(t => t.OpenRoles.OrderBy(r => r.Name))
                 .OrderBy(t => t.Name)
                 .AsNoTracking()
                 .ToList();
         }
 
-        public void TeamParticipantMatchRequest(Team team, Participant participant)
+        public bool ParticipantCanMatch(JamTeamFormingSession session, Participant participant)
         {
+            if (participant.ParticipantMatchRequest != null || participant.TeamMatchRequest != null)
+            {
+                return false;
+            }
+            if (_dbContext.Participants.Where(p => p.Session == session && p.ParticipantMatchRequest == participant).FirstOrDefault() != null)
+            {
+                return false;
+            }
+            if (_dbContext.Teams.Where(t => t.Session == session && t.ParticipantMatchRequest == participant).FirstOrDefault() != null)
+            {
+                return false;
+            }
+            return true;
+        }
+
+        public bool TeamCanMatch(JamTeamFormingSession session, Team team)
+        {
+            if (team.ParticipantMatchRequest != null)
+            {
+                return false;
+            }
+            if (_dbContext.Participants.Where(p => p.Session == session && p.TeamMatchRequest == team).FirstOrDefault() != null)
+            {
+                return false;
+            }
+            return true;
+        }
+
+        public bool TeamParticipantMatchRequest(Team team, Participant participant)
+        {
+            if (team.ParticipantMatchRequest != null
+                || participant.TeamMatchRequest != null
+                || participant.ParticipantMatchRequest != null)
+            {
+                return false;
+            }
             participant.TeamMatchRequest = team;
             _dbContext.Participants.Update(participant);
             _dbContext.SaveChanges();
+            return true;
         }
 
-        public void ParticipantTeamMatchRequest(Participant participant, Team team)
+        public bool ParticipantTeamMatchRequest(Participant participant, Team team)
         {
+            if (team.ParticipantMatchRequest != null
+                || participant.TeamMatchRequest != null
+                || participant.ParticipantMatchRequest != null)
+            {
+                return false;
+            }
             team.ParticipantMatchRequest = participant;
             _dbContext.Teams.Update(team);
             _dbContext.SaveChanges();
+            return true;
         }
 
-        public void ParticipantParticipantMatchRequest(Participant participantA, Participant participantB)
+        public bool ParticipantParticipantMatchRequest(Participant participantA, Participant participantB)
         {
+            if (participantA.TeamMatchRequest != null
+                || participantA.ParticipantMatchRequest != null
+                || participantB.TeamMatchRequest != null
+                || participantB.ParticipantMatchRequest != null)
+            {
+                return false;
+            }
             participantB.ParticipantMatchRequest = participantA;
             _dbContext.Participants.Update(participantB);
             _dbContext.SaveChanges();
+            return true;
         }
 
         public void ResolveParticipantMatch(Participant participant)
@@ -273,12 +337,36 @@ namespace JamTeamFormingTool.Services
             }
         }
 
-        public void DeleteTeam(Team team)
+        public void DeleteTeamByID(int teamID)
         {
+            Team? team = _dbContext.Teams
+                .Where(t => t.ID == teamID)
+                .Include(t => t.OpenRoles)
+                .FirstOrDefault();
+            if (team == null)
+            {
+                return;
+            }
+            team.OpenRoles.Clear();
+            _dbContext.SaveChanges();
             _dbContext.Teams.Remove(team);
             _dbContext.SaveChanges();
         }
 
-
+        public void DeleteParticipantByID(int participantID)
+        {
+            Participant? participant = _dbContext.Participants
+                .Where(p => p.ID == participantID)
+                .Include(p => p.Roles)
+                .FirstOrDefault();
+            if (participant == null)
+            {
+                return;
+            }
+            participant.Roles.Clear();
+            _dbContext.SaveChanges();
+            _dbContext.Participants.Remove(participant);
+            _dbContext.SaveChanges();
+        }
     }
 }
